Revert "主程序重构 v6 (#7)"
This reverts commit 98cf0cdc4a038296b6d11169ce43b49267dfcb8a.
