+++ README.md
@@ -24,9 +24,6 @@
     - [X]  德劳内三角定位和激光雷达定位融合
     - [x]  英雄吊射基地视角放大
     - [ ]  搬运网络和相机类至c++，通过socket通信传递信息
-    - [ ]  UI重构
-    - [ ]  EPNP独立类
-    - [ ]  回放调节
 
 - ##### 代码架构
     - 之后在写
