+++ config.py
@@ -8,53 +8,31 @@ created by 李龙 2021/12
 import os
 import numpy as np
 from real_points import real_points
-from config_type import HikCameraConfig, VideoConfig, CameraConfig, RdrConfig
+
+global global_speed
+global global_pause
+global_speed= 1.0
+global_pause = False
 
 enemy_color = 0  # 0：敌方为红色；1：敌方为蓝色
 USEABLE = {
+    "cam_left": True,
+    "cam_right": False,
     "serial": False,
     "Lidar": False,
-    "locate_state": [1, 1],
+    "locate_state": [0, 1],
 }
 BO = 0
 DEBUG = False
+using_video = True
 resource_prefix = os.path.dirname(os.path.abspath(__file__)) + "/resources/"
+record_fps = 10
 cam_config = {
-    "cam_left": {
-        "name": "cam_left",
-        "enable": True,
-        "id": "J87631625",
-        "size": (3072, 2048),
-        "roi": (0, 200, 3072, 1848),
-        "using_video": True,
-        "using_net": True,
-        "using_net_record": False,
-        "video_path": resource_prefix + "two_cam/1.avi",
-        "K_0": np.mat([[2580.7380664637653, 0.0, 1535.9830165125002],
-                       [0.0, 2582.8839945792183, 1008.784910706948],
-                       [0.0, 0.0, 1.0]]),
-        "C_0": np.mat([[-0.0640364274094021], [0.04211319930460198], [0.0010490064499735965],
-                       [-0.0003352752162304746], [0.27835581516135494]]),
-        "exposure": 15000,
-        "gain": 21,
-        "rvec": np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
-        "tvec": np.mat([[-11381.85466339], [-479.01247871], [9449.30328641]]),
-        "E_0": np.mat([
-            [0.0185759, -0.999824, 0.00251985, -0.0904854],
-            [0.0174645, -0.00219543, -0.999845, -0.132904],
-            [0.999675, 0.018617, 0.0174206, -0.421934],
-            [0, 0, 0, 1]
-        ])},
-
     "cam_right": {
-        "name": "cam_right",
-        "enable": False,
         "id": "J37877236",
         "size": (3072, 2048),
         "roi": (0, 0, 3072, 1648),
-        "using_video": True,
         "using_net": False,
-        "using_net_record": False,
         "video_path": resource_prefix + "two_cam/1.avi",
         "K_0": np.mat([[2505.2629026927225, 0.0, 1529.4286325395244],
                        [0.0, 2505.5722700649067, 1026.1378217662113],
@@ -73,106 +51,28 @@ cam_config = {
             [0, 0, 0, 1]
         ])},
 
+    "cam_left": {
+        "id": "J87631625",
+        "size": (3072, 2048),
+        "roi": (0, 200, 3072, 1848),
+        "using_net": True,
+        "video_path": resource_prefix + "two_cam/1.avi",
+        "K_0": np.mat([[2580.7380664637653, 0.0, 1535.9830165125002],
+                       [0.0, 2582.8839945792183, 1008.784910706948],
+                       [0.0, 0.0, 1.0]]),
+        "C_0": np.mat([[-0.0640364274094021], [0.04211319930460198], [0.0010490064499735965],
+                       [-0.0003352752162304746], [0.27835581516135494]]),
+        "exposure": 15000,
+        "gain": 21,
+        "rvec": np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
+        "tvec": np.mat([[-11381.85466339], [-479.01247871], [9449.30328641]]),
+        "E_0": np.mat([
+            [0.0185759, -0.999824, 0.00251985, -0.0904854],
+            [0.0174645, -0.00219543, -0.999845, -0.132904],
+            [0.999675, 0.018617, 0.0174206, -0.421934],
+            [0, 0, 0, 1]
+        ])},
 }
-
-# neo_camera_config = {
-#     "cam_left": HikCameraConfig(
-#         enable=True,
-#         roi=(0, 200, 3072, 1848),
-#         net_process=True,
-#         k_0=np.mat([[2580.7380664637653, 0.0, 1535.9830165125002],
-#                     [0.0, 2582.8839945792183, 1008.784910706948],
-#                     [0.0, 0.0, 1.0]]),
-#         c_0=np.mat([[-0.0640364274094021], [0.04211319930460198], [0.0010490064499735965],
-#                     [-0.0003352752162304746], [0.27835581516135494]]),
-#         rotate_vec=np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
-#         transform_vec=np.mat([[-11381.85466339], [-479.01247871], [9449.30328641]]),
-#         e_0=np.mat([
-#             [0.0185759, -0.999824, 0.00251985, -0.0904854],
-#             [0.0174645, -0.00219543, -0.999845, -0.132904],
-#             [0.999675, 0.018617, 0.0174206, -0.421934],
-#             [0, 0, 0, 1]
-#         ]),
-#         camera_id="J87631625",
-#         exposure=15000,
-#         gain=21,
-#     ),
-#     "cam_right": HikCameraConfig(
-#         enable=False,
-#         roi=(0, 0, 3072, 1648),
-#         net_process=False,
-#         k_0=np.mat([[2505.2629026927225, 0.0, 1529.4286325395244],
-#                     [0.0, 2505.5722700649067, 1026.1378217662113],
-#                     [0.0, 0.0, 1.0]]),
-#         c_0=np.mat([-0.06856710471358254, 0.1269396451339073,
-#                     -0.0003599605406165552, -0.0004173270419984247,
-#                     -0.141056084229664]),
-#         rotate_vec=np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
-#         transform_vec=np.mat([[-11381.85466339], [-584.01247871], [9359.30328641]]),
-#         e_0=np.mat([
-#             [0.00462803, -0.999749, 0.0219115, 0.154876],
-#             [0.0931124, -0.0213857, -0.995426, -0.0506296],
-#             [0.995645, 0.0066471, 0.0929901, -0.137197],
-#             [0, 0, 0, 1]
-#         ]),
-#         camera_id="J37877236",
-#         exposure=7000,
-#         gain=18,
-#     ),
-# }
-
-left_specs = dict(
-    k_0=np.mat([[2580.7380664637653, 0.0, 1535.9830165125002],
-                [0.0, 2582.8839945792183, 1008.784910706948],
-                [0.0, 0.0, 1.0]]),
-    c_0=np.mat([[-0.0640364274094021], [0.04211319930460198], [0.0010490064499735965],
-                [-0.0003352752162304746], [0.27835581516135494]]),
-    rotate_vec=np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
-    transform_vec=np.mat([[-11381.85466339], [-479.01247871], [9449.30328641]]),
-    e_0=np.mat([
-        [0.0185759, -0.999824, 0.00251985, -0.0904854],
-        [0.0174645, -0.00219543, -0.999845, -0.132904],
-        [0.999675, 0.018617, 0.0174206, -0.421934],
-        [0, 0, 0, 1]
-    ]))
-
-right_specs = dict(
-    k_0=np.mat([[2505.2629026927225, 0.0, 1529.4286325395244],
-                [0.0, 2505.5722700649067, 1026.1378217662113],
-                [0.0, 0.0, 1.0]]),
-    c_0=np.mat([-0.06856710471358254, 0.1269396451339073,
-                -0.0003599605406165552, -0.0004173270419984247,
-                -0.141056084229664]),
-    rotate_vec=np.mat([[1.69750257], [0.69091169], [-0.54474128]]),
-    transform_vec=np.mat([[-11381.85466339], [-584.01247871], [9359.30328641]]),
-    e_0=np.mat([
-        [0.00462803, -0.999749, 0.0219115, 0.154876],
-        [0.0931124, -0.0213857, -0.995426, -0.0506296],
-        [0.995645, 0.0066471, 0.0929901, -0.137197],
-        [0, 0, 0, 1]
-    ]))
-
-neo_camera_config = {
-    "cam_left": VideoConfig(
-        enable=True,
-        net_process=True,
-        path="/home/shiroki/radar_data/16_4_33_left.avi",
-        **left_specs
-    ),
-    # "cam_left": RdrConfig(
-    #     enable=True,
-    #     net_process=False,
-    #     endpoint="tcp://127.0.0.1:5555",
-    #     **left_specs
-    # ),
-    "cam_right": VideoConfig(
-        enable=False,
-        net_process=False,
-        path="/home/shiroki/radar_data/16_4_33_right.avi",
-        **right_specs
-    ),
-}
-
 net1_engine = resource_prefix + "net_onnx/net1.engine"
 net2_engine = resource_prefix + "net_onnx/net2.engine"
 
@@ -374,6 +274,5 @@ Delaunary_points = [
             real_points
         ]
     }
-]
 
-speed_map = [0.25, 0.5, 0.75, 1.0, 1.25, 1.5, 2.0]
+]
