--- proto/.gitignore
@@ -1,6 +1,0 @@
-*_pb2.py
-*_pb2.pyi
-
-# poetry-grpc-plugin will generate for grpc too
-*_pb2_grpc.py
-*_pb2_grpc.pyi
