+++ radar_detect/location_Delaunay.py
@@ -11,6 +11,7 @@
 import cv2 as cv
 import numpy as np
 from config import Delaunary_points, cam_config
+from mapping.drawing import draw_message
 
 
 class location_Delaunay(object):
