+++ radar_detect/location_alarm.py
@@ -8,9 +8,10 @@ import pickle
 import cv2 as cv
 import numpy as np
 import time
+import pickle as pkl
 from radar_detect.common import is_inside_polygon
-import ui.map.draw_map as draw_map  # 引入draw_map模块，使用其中的CompeteMap类
-from config import color2enemy, enemy_color, cam_config, real_size, region, test_region, choose
+import mapping.draw_map as draw_map  # 引入draw_map模块，使用其中的CompeteMap类
+from config import armor_list, color2enemy, enemy_color, cam_config, real_size, region, test_region, choose
 from radar_detect.location_Delaunay import location_Delaunay
