+++ record/protobuf/protobuf_gen.sh
@@ -1,0 +1,1 @@
+protoc --python_out=. record.proto
