--- resources/net_onnx/.gitignore
@@ -1,1 +1,0 @@
-*.engine
