--- resources/net_onnx/build.sh
@@ -1,2 +1,0 @@
-trtexec --onnx=net1.onnx --saveEngine=net1.engine
-trtexec --onnx=net2.onnx --saveEngine=net2.engine
